Re-work the code heavily and compress it, make changes
instead of selecting each button separately select all buttons with
querySelectorAll. Remove the code that added an event listener to
each button individually to populate the display and instead use the
onclick event in the HTML button tags. create the popDisplay function
that will take input from onclick event and populate the display with
it.