Remove unneeded onclick events, create new operator variables, etc
modified for loop that displays numbers to store value of num1. Modify
the old for loop used for num1 to now change background color of
clicked operator button. Added click listener to all operator buttons
to set operator variables to true. Copy the for loop used for num1
for num2 with slight modifications.